Remove snapshot-specific logic from wait_for_exec
